+++ src/pages/Admin.tsx
@@ -1,5 +1,5 @@
 import AuthenticatedNavbar from '@components/navbar/AuthenticatedNavbar';
-import { MobileBottomNavigation } from '@components/navigation/BottomNavigation';
+import BottomNavigationMobile from '@components/navbar/BottomNavigationMobile';
 import { Box, useMediaQuery, useTheme } from '@mui/material';
 
 export default function Admin() {
@@ -8,11 +8,11 @@ export default function Admin() {
 
   return (
     <>
-      {!isMobile && <AuthenticatedNavbar />}
+      <AuthenticatedNavbar />
       <Box sx={{ pb: isMobile ? 8 : 0 }}>
         <div>Admin</div>
       </Box>
-      {isMobile && <MobileBottomNavigation />}
+      <BottomNavigationMobile />
     </>
   );
 }